Migrate endpoints to dedicated namespace
