+++ .editorconfig
@@ -335,6 +335,11 @@ dotnet_diagnostic.CS1573.severity = error               # CS1573: Missing XML co
 dotnet_diagnostic.CS1591.severity = error               # CS1591: Missing XML comment for publicly visible type or member
 dotnet_diagnostic.CS1712.severity = error               # CS1712: Type parameter has no matching typeparam tag in the XML comment (but other type parameters do)
 
+[src/Immediate.Jobs.Dashboard/Endpoints/**.cs]
+
+# Dashboard internal classes
+dotnet_diagnostic.CA1812.severity = none                # CA1812: Avoid uninstantiated internal classes
+
 [tests/**.cs]
 
 dotnet_diagnostic.CA1707.severity = none                # CA1707: Identifiers should not contain underscores
