+++ src/Immediate.Jobs.Dashboard/Endpoints/CancelDashboardBatch.cs
@@ -5,21 +5,19 @@ using Immediate.Validations.Shared;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 
-#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
-
-namespace Immediate.Jobs.Dashboard;
+namespace Immediate.Jobs.Dashboard.Endpoints;
 
 [Handler]
-[MapPost("recurring/{name}/pause")]
+[MapPost("batches/{batchId}/cancel")]
 [MapGroup<DashboardApi>]
-internal static partial class PauseDashboardRecurringJob
+internal static partial class CancelDashboardBatch
 {
 	[Validate]
 	internal sealed partial record Command : IValidationTarget<Command>
 	{
 		[FromRoute]
 		[NotEmpty]
-		public required string Name { get; init; }
+		public required string BatchId { get; init; }
 	}
 
 	internal static Results<NoContent, NotFound, ProblemHttpResult> TransformResult(
@@ -30,7 +28,7 @@ internal static partial class PauseDashboardRecurringJob
 		Command command,
 		JobMonitor monitor,
 		CancellationToken cancellationToken
-	) => DashboardApiEndpointOperations.MutateRecurringAsync(
-		() => monitor.PauseRecurringAsync(command.Name, cancellationToken)
+	) => DashboardApiEndpointOperations.MutateBatchAsync(
+		() => monitor.CancelBatchAsync(command.BatchId, cancellationToken)
 	);
 }
