+++ src/Immediate.Jobs.Dashboard/Endpoints/CancelDashboardJob.cs
@@ -3,21 +3,21 @@ using Immediate.Handlers.Shared;
 using Immediate.Jobs.Shared.Apis;
 using Immediate.Validations.Shared;
 using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.AspNetCore.Mvc;
 
-#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
-
-namespace Immediate.Jobs.Dashboard;
+namespace Immediate.Jobs.Dashboard.Endpoints;
 
 [Handler]
-[MapDelete("batches/{batchId}")]
+[MapPost("jobs/{jobId}/cancel")]
 [MapGroup<DashboardApi>]
-internal static partial class DeleteDashboardBatch
+internal static partial class CancelDashboardJob
 {
 	[Validate]
 	internal sealed partial record Command : IValidationTarget<Command>
 	{
+		[FromRoute]
 		[NotEmpty]
-		public required string BatchId { get; init; }
+		public required string JobId { get; init; }
 	}
 
 	internal static Results<NoContent, NotFound, ProblemHttpResult> TransformResult(
@@ -28,7 +28,7 @@ internal static partial class DeleteDashboardBatch
 		Command command,
 		JobMonitor monitor,
 		CancellationToken cancellationToken
-	) => DashboardApiEndpointOperations.MutateBatchAsync(
-		() => monitor.DeleteBatchAsync(command.BatchId, cancellationToken)
+	) => DashboardApiEndpointOperations.MutateJobAsync(
+		() => monitor.CancelJobAsync(command.JobId, cancellationToken)
 	);
 }
