+++ src/Immediate.Jobs.Dashboard/Endpoints/DashboardApi.cs
@@ -2,11 +2,9 @@ using Immediate.Apis.Shared;
 using Immediate.Handlers.Shared;
 using Immediate.Validations.Shared;
 
-#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
-
 [assembly: Behaviors(typeof(ValidationBehavior<,>))]
 
-namespace Immediate.Jobs.Dashboard;
+namespace Immediate.Jobs.Dashboard.Endpoints;
 
 [RouteGroup("api")]
 internal sealed partial class DashboardApi;
