+++ src/Immediate.Jobs.Dashboard/Endpoints/GetDashboardJobTelemetryLinks.cs
@@ -5,23 +5,18 @@ using Immediate.Validations.Shared;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.Extensions.Options;
 
-#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
-
-namespace Immediate.Jobs.Dashboard;
+namespace Immediate.Jobs.Dashboard.Endpoints;
 
 [Handler]
-[MapGet("jobs/{jobId}/executions/{executionNumber:int}/telemetry-links")]
+[MapGet("jobs/{jobId}/telemetry-links")]
 [MapGroup<DashboardApi>]
-internal static partial class GetDashboardJobExecutionTelemetryLinks
+internal static partial class GetDashboardJobTelemetryLinks
 {
 	[Validate]
 	internal sealed partial record Query : IValidationTarget<Query>
 	{
 		[NotEmpty]
 		public required string JobId { get; init; }
-
-		[GreaterThan(0)]
-		public required int ExecutionNumber { get; init; }
 	}
 
 	internal static Results<JsonHttpResult<IReadOnlyList<JobTelemetryLink>>, NotFound> TransformResult(
@@ -35,7 +30,7 @@ internal static partial class GetDashboardJobExecutionTelemetryLinks
 		CancellationToken cancellationToken
 	) => DashboardApiEndpointOperations.GetJobTelemetryLinksAsync(
 		query.JobId,
-		query.ExecutionNumber,
+		executionNumber: null,
 		monitor,
 		options.Value,
 		cancellationToken
