+++ src/Immediate.Jobs.Dashboard/Endpoints/StreamDashboardBatch.cs
@@ -1,25 +1,29 @@
 using Immediate.Apis.Shared;
 using Immediate.Handlers.Shared;
 using Immediate.Jobs.Shared.Apis;
+using Immediate.Validations.Shared;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.Extensions.Options;
 
-#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
-
-namespace Immediate.Jobs.Dashboard;
+namespace Immediate.Jobs.Dashboard.Endpoints;
 
 [Handler]
-[MapGet("events")]
+[MapGet("batches/{batchId}/stream")]
 [MapGroup<DashboardApi>]
-internal static partial class StreamDashboardEvents
+internal static partial class StreamDashboardBatch
 {
-	internal sealed record Query;
+	[Validate]
+	internal sealed partial record Query : IValidationTarget<Query>
+	{
+		[NotEmpty]
+		public required string BatchId { get; init; }
+	}
 
 	internal static EmptyHttpResult TransformResult() => TypedResults.Empty;
 
 	private static async ValueTask HandleAsync(
-		Query _,
+		Query query,
 		IHttpContextAccessor httpContextAccessor,
 		JobMonitor monitor,
 		TimeProvider timeProvider,
@@ -27,8 +31,9 @@ internal static partial class StreamDashboardEvents
 		CancellationToken cancellationToken
 	)
 	{
-		await DashboardApiEndpointOperations.StreamEventsAsync(
+		await DashboardApiEndpointOperations.StreamBatchEventsAsync(
 			httpContextAccessor.HttpContext ?? throw new InvalidOperationException("No active dashboard HTTP request was found."),
+			query.BatchId,
 			monitor,
 			timeProvider,
 			options.Value.UpdateInterval,
