+++ src/Immediate.Jobs.Dashboard/Immediate.Jobs.Dashboard.csproj
@@ -47,35 +47,22 @@
     <ProjectReference Include="../Immediate.Jobs/Immediate.Jobs.csproj" ReferenceOutputAssembly="false" />
   </ItemGroup>
 
-  <Target Name="RestoreDashboardClient"
-	Inputs="$(DashboardClientDirectory)/package.json;$(DashboardClientDirectory)/package-lock.json"
-	Outputs="$(DashboardClientDirectory)/node_modules/.package-lock.json">
+  <Target Name="RestoreDashboardClient" Inputs="$(DashboardClientDirectory)/package.json;$(DashboardClientDirectory)/package-lock.json" Outputs="$(DashboardClientDirectory)/node_modules/.package-lock.json">
 	<Exec Command="npm ci" WorkingDirectory="$(DashboardClientDirectory)" />
   </Target>
 
-  <Target Name="BuildDashboardClient"
-	BeforeTargets="DispatchToInnerBuilds;PrepareForBuild"
-	DependsOnTargets="RestoreDashboardClient"
-	Inputs="@(DashboardClientSource)"
-	Outputs="$(DashboardBuildStamp);$(DashboardAssetsDirectory)/app.js;$(DashboardAssetsDirectory)/app.css;$(DashboardAssetsDirectory)/index.html"
-	Condition="'$(DesignTimeBuild)' != 'true'">
+  <Target Name="BuildDashboardClient" BeforeTargets="DispatchToInnerBuilds;PrepareForBuild" DependsOnTargets="RestoreDashboardClient" Inputs="@(DashboardClientSource)" Outputs="$(DashboardBuildStamp);$(DashboardAssetsDirectory)/app.js;$(DashboardAssetsDirectory)/app.css;$(DashboardAssetsDirectory)/index.html" Condition="'$(DesignTimeBuild)' != 'true'">
 	<Exec Command="npm run build" WorkingDirectory="$(DashboardClientDirectory)" />
 	<Exec Command="npm run size" WorkingDirectory="$(DashboardClientDirectory)" />
 	<MakeDir Directories="$(BaseIntermediateOutputPath)" />
 	<Touch Files="$(DashboardBuildStamp)" AlwaysCreate="true" />
   </Target>
 
-  <Target Name="IncludeDashboardAssets"
-	BeforeTargets="AssignTargetPaths"
-	DependsOnTargets="BuildDashboardClient"
-	Condition="'$(IsCrossTargetingBuild)' != 'true'">
+  <Target Name="IncludeDashboardAssets" BeforeTargets="AssignTargetPaths" DependsOnTargets="BuildDashboardClient" Condition="'$(IsCrossTargetingBuild)' != 'true'">
 	<ItemGroup>
-	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/app.js"
-		LogicalName="$(RootNamespace).Assets.app.js" />
-	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/app.css"
-		LogicalName="$(RootNamespace).Assets.app.css" />
-	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/index.html"
-		LogicalName="$(RootNamespace).Assets.index.html" />
+	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/app.js" LogicalName="$(RootNamespace).Assets.app.js" />
+	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/app.css" LogicalName="$(RootNamespace).Assets.app.css" />
+	  <EmbeddedResource Include="$(DashboardAssetsDirectory)/index.html" LogicalName="$(RootNamespace).Assets.index.html" />
 	</ItemGroup>
   </Target>
