+++ src/Immediate.Jobs.Dashboard/ImmediateJobsDashboardEndpointRouteBuilderExtensions.cs
@@ -28,6 +28,7 @@ public static class ImmediateJobsDashboardEndpointRouteBuilderExtensions
 	{
 		ArgumentNullException.ThrowIfNull(endpoints);
 		ArgumentException.ThrowIfNullOrWhiteSpace(prefix);
+
 		if (prefix[0] != '/')
 			prefix = "/" + prefix;
 
@@ -49,18 +50,31 @@ public static class ImmediateJobsDashboardEndpointRouteBuilderExtensions
 		_ = group.AddEndpointFilter(new DashboardValidationFilter());
 
 		_ = group.MapImmediateJobsDashboardEndpoints();
-		_ = group.MapGet("/", (Delegate)((HttpContext context) =>
-			context.Request.Path.Value is { Length: > 0 } path && path[^1] == '/'
-				? DashboardAssets.GetIndexAsync(context, prefix)
-				: Task.FromResult(Results.Redirect(prefix + "/"))
-		)).ExcludeFromDescription();
+
+		_ = group
+			.MapGet(
+				"/",
+				(Delegate)(
+					(HttpContext context) =>
+						context.Request.Path.Value is { Length: > 0 } path && path[^1] == '/'
+							? DashboardAssets.GetIndexAsync(context, prefix)
+							: Task.FromResult(Results.Redirect(prefix + "/"))
+				)
+			)
+			.ExcludeFromDescription();
+
 		_ = group.MapGet("/app.css", () => DashboardAssets.GetAsync("app.css")).ExcludeFromDescription();
 		_ = group.MapGet("/app.js", () => DashboardAssets.GetAsync("app.js")).ExcludeFromDescription();
-		_ = group.MapGet("/{**path}", (string path, HttpContext context) =>
-			path.StartsWith("api/", StringComparison.OrdinalIgnoreCase)
-				? Task.FromResult(Results.NotFound())
-				: DashboardAssets.GetIndexAsync(context, prefix)
-		).WithOrder(int.MaxValue).ExcludeFromDescription();
+
+		_ = group.MapGet(
+			"/{**path}",
+			(string path, HttpContext context) =>
+				path.StartsWith("api/", StringComparison.OrdinalIgnoreCase)
+					? Task.FromResult(Results.NotFound())
+					: DashboardAssets.GetIndexAsync(context, prefix)
+			)
+			.WithOrder(int.MaxValue)
+			.ExcludeFromDescription();
 
 		return group;
 	}
